updated modules fileto add genre
