+++ schema.sql
@@ -14,20 +14,14 @@ CREATE TABLE Books (
   FOREIGN KEY (label_id) REFERENCES Labels(id)
 )
 
-<<<<<<< HEAD
 CREATE TABLE Labels (
   id GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   title VARCHAR(30) NOT NULL,
   color VARCHAR(20)
 )
-=======
->>>>>>> 58bce64cea15238305b5f66dc0d80386c21f28c7
 CREATE TABLE music_album (
     id int GENERATED BY DEFAULT AS IDENTITY,
-    genre_id int,
-    author_id int,
-    source_id int,
-    label_id int,
+    genre_id int
     publish_date date,
     archived boolean,
     on_spotify boolean,
@@ -40,8 +34,5 @@ CREATE TABLE genre (
     title varchar(50),
     genre varchar(50),
     CONSTRAINT genre_pk PRIMARY KEY (id)
-<<<<<<< HEAD
 )
-=======
 )
->>>>>>> 58bce64cea15238305b5f66dc0d80386c21f28c7
